retrieval: wip improved retrieval commands
