+++ cli/client.go
@@ -1028,225 +1028,6 @@ var clientFindCmd = &cli.Command{
 	},
 }
 
-const DefaultMaxRetrievePrice = "0.01"
-
-var clientRetrieveCmd = &cli.Command{
-	Name:      "retrieve",
-	Usage:     "Retrieve data from network",
-	ArgsUsage: "[dataCid outputPath]",
-	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:  "from",
-			Usage: "address to send transactions from",
-		},
-		&cli.BoolFlag{
-			Name:  "car",
-			Usage: "export to a car file instead of a regular file",
-		},
-		&cli.StringFlag{
-			Name:  "miner",
-			Usage: "miner address for retrieval, if not present it'll use local discovery",
-		},
-		&cli.StringFlag{
-			Name:  "datamodel-path-selector",
-			Usage: "a rudimentary (DM-level-only) text-path selector, allowing for sub-selection within a deal",
-		},
-		&cli.StringFlag{
-			Name:  "maxPrice",
-			Usage: fmt.Sprintf("maximum price the client is willing to consider (default: %s FIL)", DefaultMaxRetrievePrice),
-		},
-		&cli.StringFlag{
-			Name:  "pieceCid",
-			Usage: "require data to be retrieved from a specific Piece CID",
-		},
-		&cli.BoolFlag{
-			Name: "allow-local",
-		},
-	},
-	Action: func(cctx *cli.Context) error {
-		if cctx.NArg() != 2 {
-			return ShowHelp(cctx, fmt.Errorf("incorrect number of arguments"))
-		}
-
-		fapi, closer, err := GetFullNodeAPIV1(cctx)
-		if err != nil {
-			return err
-		}
-		defer closer()
-		ctx := ReqContext(cctx)
-		afmt := NewAppFmt(cctx.App)
-
-		var payer address.Address
-		if cctx.String("from") != "" {
-			payer, err = address.NewFromString(cctx.String("from"))
-		} else {
-			payer, err = fapi.WalletDefaultAddress(ctx)
-		}
-		if err != nil {
-			return err
-		}
-
-		file, err := cid.Parse(cctx.Args().Get(0))
-		if err != nil {
-			return err
-		}
-
-		var pieceCid *cid.Cid
-		if cctx.String("pieceCid") != "" {
-			parsed, err := cid.Parse(cctx.String("pieceCid"))
-			if err != nil {
-				return err
-			}
-			pieceCid = &parsed
-		}
-
-		var eref *lapi.ExportRef
-		if cctx.Bool("allow-local") {
-			imports, err := fapi.ClientListImports(ctx)
-			if err != nil {
-				return err
-			}
-
-			for _, i := range imports {
-				if i.Root != nil && i.Root.Equals(file) {
-					eref = &lapi.ExportRef{
-						Root:         file,
-						FromLocalCAR: i.CARPath,
-					}
-					break
-				}
-			}
-		}
-
-		// no local found, so make a retrieval
-		if eref == nil {
-			var offer api.QueryOffer
-			minerStrAddr := cctx.String("miner")
-			if minerStrAddr == "" { // Local discovery
-				offers, err := fapi.ClientFindData(ctx, file, pieceCid)
-
-				var cleaned []api.QueryOffer
-				// filter out offers that errored
-				for _, o := range offers {
-					if o.Err == "" {
-						cleaned = append(cleaned, o)
-					}
-				}
-
-				offers = cleaned
-
-				// sort by price low to high
-				sort.Slice(offers, func(i, j int) bool {
-					return offers[i].MinPrice.LessThan(offers[j].MinPrice)
-				})
-				if err != nil {
-					return err
-				}
-
-				// TODO: parse offer strings from `client find`, make this smarter
-				if len(offers) < 1 {
-					fmt.Println("Failed to find file")
-					return nil
-				}
-				offer = offers[0]
-			} else { // Directed retrieval
-				minerAddr, err := address.NewFromString(minerStrAddr)
-				if err != nil {
-					return err
-				}
-				offer, err = fapi.ClientMinerQueryOffer(ctx, minerAddr, file, pieceCid)
-				if err != nil {
-					return err
-				}
-			}
-			if offer.Err != "" {
-				return fmt.Errorf("offer error: %s", offer.Err)
-			}
-
-			maxPrice := types.MustParseFIL(DefaultMaxRetrievePrice)
-
-			if cctx.String("maxPrice") != "" {
-				maxPrice, err = types.ParseFIL(cctx.String("maxPrice"))
-				if err != nil {
-					return xerrors.Errorf("parsing maxPrice: %w", err)
-				}
-			}
-
-			if offer.MinPrice.GreaterThan(big.Int(maxPrice)) {
-				return xerrors.Errorf("failed to find offer satisfying maxPrice: %s", maxPrice)
-			}
-
-			o := offer.Order(payer)
-
-			subscribeEvents, err := fapi.ClientGetRetrievalUpdates(ctx)
-			if err != nil {
-				return xerrors.Errorf("error setting up retrieval updates: %w", err)
-			}
-			retrievalRes, err := fapi.ClientRetrieve(ctx, o)
-			if err != nil {
-				return xerrors.Errorf("error setting up retrieval: %w", err)
-			}
-
-		readEvents:
-			for {
-				var evt api.RetrievalInfo
-				select {
-				case <-ctx.Done():
-					return xerrors.New("Retrieval Timed Out")
-				case evt = <-subscribeEvents:
-					if evt.ID != retrievalRes.DealID {
-						// we can't check the deal ID ahead of time because:
-						// 1. We need to subscribe before retrieving.
-						// 2. We won't know the deal ID until after retrieving.
-						continue
-					}
-				}
-
-				event := "New"
-				if evt.Event != nil {
-					event = retrievalmarket.ClientEvents[*evt.Event]
-				}
-
-				afmt.Printf("> Recv: %s, Paid %s, %s (%s)\n",
-					types.SizeStr(types.NewInt(evt.BytesReceived)),
-					types.FIL(evt.TotalPaid),
-					event,
-					retrievalmarket.DealStatuses[evt.Status],
-				)
-				switch evt.Status {
-				case retrievalmarket.DealStatusCompleted:
-					break readEvents
-				case retrievalmarket.DealStatusRejected:
-					return xerrors.Errorf("Retrieval Proposal Rejected: %s", evt.Message)
-				case
-					retrievalmarket.DealStatusDealNotFound,
-					retrievalmarket.DealStatusErrored:
-					return xerrors.Errorf("Retrieval Error: %s", evt.Message)
-				}
-			}
-
-			eref = &lapi.ExportRef{
-				Root:   file,
-				DealID: retrievalRes.DealID,
-			}
-		}
-
-		if sel := api.Selector(cctx.String("datamodel-path-selector")); sel != "" {
-			eref.DAGs = append(eref.DAGs, api.DagSpec{DataSelector: &sel})
-		}
-
-		err = fapi.ClientExport(ctx, *eref, lapi.FileRef{
-			Path:  cctx.Args().Get(1),
-			IsCAR: cctx.Bool("car"),
-		})
-		if err != nil {
-			return err
-		}
-		afmt.Println("Success")
-		return nil
-	},
-}
-
 var clientListRetrievalsCmd = &cli.Command{
 	Name:  "list-retrievals",
 	Usage: "List retrieval market deals",
